+++ packages/manager/src/features/Images/utils.ts
@@ -1,0 +1,20 @@
+import { getGrantData, getProfileData } from 'src/queries/profile';
+
+export const useImageAndLinodeGrantCheck = () => {
+  const profile = getProfileData();
+  const grants = getGrantData();
+
+  const canCreateImage =
+    Boolean(!profile?.restricted) || Boolean(grants?.global?.add_images);
+
+  // Unrestricted users can create Images from any disk;
+  // Restricted users need read_write on the Linode they're trying to Imagize
+  // (in addition to the global add_images grant).
+  const permissionedLinodes = profile?.restricted
+    ? grants?.linode
+        .filter((thisGrant) => thisGrant.permissions === 'read_write')
+        .map((thisGrant) => thisGrant.id) ?? []
+    : null;
+
+  return { canCreateImage, permissionedLinodes };
+};
