+++ packages/manager/src/features/linodes/LinodesDetail/LinodeRebuild/RebuildFromImage.test.tsx
@@ -6,12 +6,6 @@ import { CombinedProps, RebuildFromImage } from './RebuildFromImage';
 
 jest.mock('src/utilities/scrollErrorIntoView');
 jest.mock('src/components/EnhancedSelect/Select');
-jest.mock('src/hooks/useReduxLoad', () => ({
-  useReduxLoad: () => jest.fn().mockReturnValue({ _loading: false }),
-}));
-jest.mock('src/hooks/useImages', () => ({
-  useImages: jest.fn().mockResolvedValue({ error: {} }),
-}));
 
 const props: CombinedProps = {
   linodeId: 1234,
